refactor(domain/board/Position): 경로에 기물이 있는지 확인하는 기능을 Position 으로 책임으로 변경
