--- src/main/java/chess/domain/board/PathCheck.java
@@ -1,14 +1,0 @@
-package chess.domain.board;
-
-import java.util.List;
-import java.util.function.Predicate;
-
-public class PathCheck {
-	public static boolean check(Position beforePosition,
-		Position afterPosition,
-		Predicate<Position> positionPredicate) {
-		List<Position> path = beforePosition.pathTo(afterPosition);
-		return path.stream()
-			.allMatch(positionPredicate);
-	}
-}
