+++ src/test/java/chess/domain/board/PositionTest.java
@@ -2,6 +2,8 @@ package chess.domain.board;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import java.util.Objects;
+
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -91,5 +93,38 @@ public class PositionTest {
 		//when
 		//then
 		assertThat(H1).isGreaterThan(B2);
+	}
+
+	@DisplayName("a1에서 a3 사이에는 a2가 있다.")
+	@Test
+	void path_a1_a3() {
+		//given
+		Position a1 = new Position(Column.A, Row.ONE);
+		Position a3 = new Position(Column.A, Row.THREE);
+		//then
+		assertThat(a1.existObstacleToOtherPosition(a3,
+			(position -> Objects.equals(position, new Position(Column.A, Row.TWO))))).isTrue();
+	}
+
+	@DisplayName("a1에서 c1 사이에는 b1이 있다.")
+	@Test
+	void path_a1_c1() {
+		//given
+		Position a1 = new Position(Column.A, Row.ONE);
+		Position c1 = new Position(Column.C, Row.ONE);
+		//then
+		assertThat(a1.existObstacleToOtherPosition(c1,
+			(position -> Objects.equals(position, new Position(Column.B, Row.ONE))))).isTrue();
+	}
+
+	@DisplayName("e1에서 c3 사이에는 d2가 있다.")
+	@Test
+	void path_e1_c3() {
+		//given
+		Position e1 = new Position(Column.E, Row.ONE);
+		Position c3 = new Position(Column.C, Row.THREE);
+		//then
+		assertThat(e1.existObstacleToOtherPosition(c3,
+			(position -> Objects.equals(position, new Position(Column.D, Row.TWO))))).isTrue();
 	}
 }
